Move mocks stuff to mocks package.
