+++ st2tests/st2tests/mocks/sensor.py
